requirements.txt
openai>=1.0.0 
azure-search-documents>=11.6.0 
python-dotenv>=1.0.0
fastapi
uvicorn
